+++ examples/app/main.go
@@ -27,7 +27,7 @@ var consumerDeviceCapabilities = mediasoup.RtpCapabilities{
 			Parameters: mediasoup.RtpCodecSpecificParameters{
 				RtpParameter: h264.RtpParameter{
 					LevelAsymmetryAllowed: 1,
-					PacketizationMode:     1,
+					PacketizationMode:     Uint8(1),
 					ProfileLevelId:        "4d0032",
 				},
 			},
@@ -89,7 +89,7 @@ func main() {
 				Parameters: mediasoup.RtpCodecSpecificParameters{
 					RtpParameter: h264.RtpParameter{
 						LevelAsymmetryAllowed: 1,
-						PacketizationMode:     1,
+						PacketizationMode:     Uint8(1),
 						ProfileLevelId:        "4d0032",
 					},
 				},
@@ -120,7 +120,7 @@ func main() {
 					ClockRate:   90000,
 					Parameters: mediasoup.RtpCodecSpecificParameters{
 						RtpParameter: h264.RtpParameter{
-							PacketizationMode: 1,
+							PacketizationMode: Uint8(1),
 							ProfileLevelId:    "4d0032",
 						},
 					},
